+++ src/lua/builtin_kinds.lua
@@ -1,9 +1,0 @@
-tpane.kind {
-  name = "pane",
-  detect = function(_p)
-    return true
-  end,
-  label = function(p)
-    return p.command
-  end,
-}